Applying design patterns and access check

See merge request cse2115/2022-2023/SEM06b!42

+++ store-microservice/src/main/java/nl/tudelft/sem/group06b/store/api/StoreController.java
@@ -1,8 +1,10 @@
 package nl.tudelft.sem.group06b.store.api;
 
 import java.util.List;
+import java.util.Locale;
+import nl.tudelft.sem.group06b.store.authentication.AuthManager;
 import nl.tudelft.sem.group06b.store.domain.Location;
-import nl.tudelft.sem.group06b.store.domain.Store;
+import nl.tudelft.sem.group06b.store.domain.store.Store;
 import nl.tudelft.sem.group06b.store.model.ModifyStoreRequestModel;
 import nl.tudelft.sem.group06b.store.service.StoreService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,6 +26,10 @@ public class StoreController {
 
     private final transient StoreService storeService;
 
+    private final transient AuthManager authManager;
+
+    private final transient String regionalManager = "regional_manager";
+
     private transient Long storeId;
 
 
@@ -33,7 +39,8 @@ public class StoreController {
      * @param storeService The store service.
      */
     @Autowired
-    public StoreController(StoreService storeService) {
+    public StoreController(AuthManager authManager, StoreService storeService) {
+        this.authManager = authManager;
         this.storeService = storeService;
     }
 
@@ -44,7 +51,10 @@ public class StoreController {
      */
     @GetMapping("/showAllStores")
     public ResponseEntity<List<Store>> queryAllStores() {
-        return ResponseEntity.ok(storeService.queryAllStores());
+        if (authManager.getRole().toLowerCase(Locale.ROOT).equals(regionalManager)) {
+            return ResponseEntity.ok(storeService.queryAllStores());
+        }
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Only regional managers can view all stores");
     }
 
     /**
@@ -55,11 +65,14 @@ public class StoreController {
      */
     @PostMapping("/addStore")
     public ResponseEntity<String> addStore(@RequestBody ModifyStoreRequestModel request) {
-
+        if (!authManager.getRole().toLowerCase(Locale.ROOT).equals(regionalManager)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Only regional managers can add stores");
+        }
         try {
             String name = request.getName();
             Location location = new Location(request.getStoreLocation());
-            storeService.addStore(name, location);
+            String manager = request.getManager();
+            storeService.addStore(name, location, manager);
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
         }
@@ -74,7 +87,9 @@ public class StoreController {
      */
     @DeleteMapping("/removeStore/{storeId}")
     public ResponseEntity<String> removeStore(@PathVariable Long storeId) {
-
+        if (!authManager.getRole().toLowerCase(Locale.ROOT).equals(regionalManager)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Only regional managers can remove stores");
+        }
         try {
             storeService.removeStore(storeId);
         } catch (Exception e) {
@@ -96,6 +111,17 @@ public class StoreController {
     }
 
     /**
+     * Validates if the store manager exists.
+     *
+     * @param manager memeberId of the manager.
+     * @return True if the given store location exists.
+     */
+    @GetMapping("/validateManager/{manager}")
+    public ResponseEntity<Boolean> validateManager(@PathVariable String manager) {
+        return ResponseEntity.ok(storeService.validateManager(manager));
+    }
+
+    /**
      * Gets the store id from the database.
      *
      * @param address address of the store.
@@ -103,9 +129,35 @@ public class StoreController {
      */
     @GetMapping("/getStoreId/{address}")
     public ResponseEntity<Long> getStoreId(@PathVariable String address) {
+        if (!storeService.validateManager(authManager.getMemberId())
+                && !authManager.getRole().toLowerCase(Locale.ROOT).equals(regionalManager)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Only store managers and regional managers can get store Ids");
+        }
         try {
             Location location = new Location(address);
             storeId = storeService.retrieveStoreId(location);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+        return ResponseEntity.ok(storeId);
+    }
+
+    /**
+     * Gets the store id from the database.
+     *
+     * @param manager address of the store.
+     * @return The store id.
+     */
+    @GetMapping("/getStoreIdManager/{manager}")
+    public ResponseEntity<Long> getStoreIdFromManager(@PathVariable String manager) {
+        if (!storeService.validateManager(authManager.getMemberId())
+                && !authManager.getRole().toLowerCase(Locale.ROOT).equals(regionalManager)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Only store managers and regional managers can get store Ids");
+        }
+        try {
+            storeId = storeService.retrieveStoreId(manager);
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
         }
